fix fix fix fix

+++ internal/api/handlers/courses/handlers.go
@@ -133,34 +133,39 @@ func CreateCourseHandler(c *components.HTTPComponents) {
 		return
 	}
 
-	thumbnailPictureFile, err := helpers.ConvertBase64ImageToFile(courseRequest.ThumbnailURL)
-	if err != nil {
-		components.HttpErrorResponse(c, http.StatusInternalServerError, errutil.ErrUnexpectedError)
-		return
+	thumbnailURL := ""
+	if courseRequest.ThumbnailURL == "" {
+		thumbnailPictureFile, err := helpers.ConvertBase64ImageToFile(courseRequest.ThumbnailURL)
+		if err != nil {
+			components.HttpErrorResponse(c, http.StatusInternalServerError, errutil.ErrUnexpectedError)
+			return
+		}
+		defer os.Remove(thumbnailPictureFile.Name())
+
+		croppedPictureFile, err := helpers.ResizeImage(thumbnailPictureFile, 1280, 720)
+		if err != nil {
+			log.Println("Error resizing image", err)
+
+			components.HttpErrorResponse(c, http.StatusInternalServerError, errutil.ErrUnexpectedError)
+			return
+		}
+		defer os.Remove(croppedPictureFile.Name())
+
+		thumbnailURL = fmt.Sprintf("thumbnails/%s", croppedPictureFile.Name())
+		s3Client := aws.GetS3Client(aws.GetAWSConfig(c.Components))
+		err = s3Client.UploadFile(c.Components.Settings.String("aws.coursesBucketName"), thumbnailURL, croppedPictureFile)
+		if err != nil {
+			log.Println("Error uploading file to S3", err)
+
+			components.HttpErrorResponse(c, http.StatusInternalServerError, errutil.ErrUnexpectedError)
+			return
+		}
+
+		thumbnailURL = c.Components.Settings.String("aws.coursesBucketURL") + thumbnailURL
 	}
-	defer os.Remove(thumbnailPictureFile.Name())
 
-	croppedPictureFile, err := helpers.ResizeImage(thumbnailPictureFile, 1280, 720)
-	if err != nil {
-		log.Println("Error resizing image", err)
-
-		components.HttpErrorResponse(c, http.StatusInternalServerError, errutil.ErrUnexpectedError)
-		return
-	}
-	defer os.Remove(croppedPictureFile.Name())
-
-	thumbnailURL := fmt.Sprintf("thumbnails/%s", croppedPictureFile.Name())
-	s3Client := aws.GetS3Client(aws.GetAWSConfig(c.Components))
-	err = s3Client.UploadFile(c.Components.Settings.String("aws.coursesBucketName"), thumbnailURL, croppedPictureFile)
-	if err != nil {
-		log.Println("Error uploading file to S3", err)
-
-		components.HttpErrorResponse(c, http.StatusInternalServerError, errutil.ErrUnexpectedError)
-		return
-	}
-
-	courseRequest.ThumbnailURL = c.Components.Settings.String("aws.coursesBucketURL") + thumbnailURL
 	course := courseRequest.ToEntity()
+	courseRequest.ThumbnailURL = thumbnailURL
 
 	err = c.Components.Resources.Courses.Create(course)
 
@@ -238,41 +243,40 @@ func UpdateCourseHandler(c *components.HTTPComponents) {
 		return
 	}
 
-	thumbnailPictureFile, err := helpers.ConvertBase64ImageToFile(courseRequest.ThumbnailURL)
-	if err != nil {
-		components.HttpErrorResponse(c, http.StatusInternalServerError, errutil.ErrUnexpectedError)
-		return
+	thumbnailURL := ""
+	if courseRequest.ThumbnailURL == "" {
+		thumbnailPictureFile, err := helpers.ConvertBase64ImageToFile(courseRequest.ThumbnailURL)
+		if err != nil {
+			components.HttpErrorResponse(c, http.StatusInternalServerError, errutil.ErrUnexpectedError)
+			return
+		}
+		defer os.Remove(thumbnailPictureFile.Name())
+
+		croppedPictureFile, err := helpers.ResizeImage(thumbnailPictureFile, 1280, 720)
+		if err != nil {
+			log.Println("Error resizing image", err)
+
+			components.HttpErrorResponse(c, http.StatusInternalServerError, errutil.ErrUnexpectedError)
+			return
+		}
+		defer os.Remove(croppedPictureFile.Name())
+
+		thumbnailURL = fmt.Sprintf("thumbnails/%s", croppedPictureFile.Name())
+		s3Client := aws.GetS3Client(aws.GetAWSConfig(c.Components))
+		err = s3Client.UploadFile(c.Components.Settings.String("aws.coursesBucketName"), thumbnailURL, croppedPictureFile)
+		if err != nil {
+			log.Println("Error uploading file to S3", err)
+
+			components.HttpErrorResponse(c, http.StatusInternalServerError, errutil.ErrUnexpectedError)
+			return
+		}
+
+		thumbnailURL = c.Components.Settings.String("aws.coursesBucketURL") + thumbnailURL
 	}
-	defer os.Remove(thumbnailPictureFile.Name())
-
-	croppedPictureFile, err := helpers.ResizeImage(thumbnailPictureFile, 1280, 720)
-	if err != nil {
-		log.Println("Error resizing image", err)
-
-		components.HttpErrorResponse(c, http.StatusInternalServerError, errutil.ErrUnexpectedError)
-		return
-	}
-	defer os.Remove(croppedPictureFile.Name())
-
-	thumbnailURL := fmt.Sprintf("thumbnails/%s", croppedPictureFile.Name())
-	s3Client := aws.GetS3Client(aws.GetAWSConfig(c.Components))
-	err = s3Client.UploadFile(c.Components.Settings.String("aws.coursesBucketName"), thumbnailURL, croppedPictureFile)
-	if err != nil {
-		log.Println("Error uploading file to S3", err)
-
-		components.HttpErrorResponse(c, http.StatusInternalServerError, errutil.ErrUnexpectedError)
-		return
-	}
-
-	err = os.Remove(thumbnailPictureFile.Name())
-	if err != nil {
-		log.Println("Error removing file", err)
-	}
-
-	courseRequest.ThumbnailURL = c.Components.Settings.String("aws.coursesBucketURL") + thumbnailURL
 
 	course := courseRequest.ToEntity()
 	course.ID = uuid.MustParse(id)
+	courseRequest.ThumbnailURL = thumbnailURL
 
 	rowsAffected, err := c.Components.Resources.Courses.Update(course)
